refactor(upstream): extract pool/tools.rs + pool/tools_call.rs (lab-kvji.12.8 step 15)
- tools.rs: tool queries (healthy_tools*, find_tool*, subject_scoped_tools,
  routable_upstream_names, tool_schema, tool_exposure_rows, cached_upstream_summary,
  upstream_runtime_metadata, upstream_tool_health) + 2 hidden_upstream_tools tests.
  pub(super) FLIP: has_healthy_tools_for_upstream (ensure.rs calls it).
- tools_call.rs: subject_scoped_call_tool + call_tool + call_tool_times_out test.
  NO-TOUCH: subject_scoped_call_tool subject threading preserved (verified). Uses
  acquire_peer + UpstreamRequestLog ctors (pub(super) from steps 7/3).
Behavior unchanged. 1423 green. redact count unchanged (4).

Bead: lab-kvji.12.8
