+++ crates/lab/src/dispatch/upstream/pool.rs
@@ -12,8 +12,8 @@ use futures::future::BoxFuture;
 use futures::stream::FuturesUnordered;
 use rmcp::RoleClient;
 use rmcp::model::{
-    AnnotateAble, CallToolRequestParams, CallToolResult, GetPromptRequestParams, GetPromptResult,
-    Prompt, RawResource, ReadResourceResult, Resource,
+    AnnotateAble, GetPromptRequestParams, GetPromptResult, Prompt, RawResource, ReadResourceResult,
+    Resource,
 };
 use serde_json::Value;
 use tokio::sync::{Mutex, RwLock};
@@ -24,8 +24,7 @@ use crate::oauth::upstream::cache::OauthClientCache;
 use crate::registry::RegisteredService;
 
 use super::types::{
-    UpstreamCapability, UpstreamEntry, UpstreamHealth, UpstreamRuntimeMetadata,
-    UpstreamRuntimeOwner, UpstreamTool, UpstreamToolExposureRow,
+    UpstreamCapability, UpstreamEntry, UpstreamRuntimeMetadata, UpstreamRuntimeOwner,
 };
 
 mod capability;
@@ -43,6 +42,8 @@ mod probe;
 mod registration;
 #[cfg(test)]
 mod testsupport;
+mod tools;
+mod tools_call;
 mod validate;
 
 pub(crate) use helpers::redact_resource_uri_for_logging;
@@ -162,344 +163,6 @@ impl UpstreamPool {
         self
     }
 
-    /// Get all healthy upstream tools.
-    pub async fn healthy_tools(&self) -> Vec<UpstreamTool> {
-        let catalog = self.catalog.read().await;
-        catalog
-            .values()
-            .filter(|entry| entry.tool_health.is_routable())
-            .flat_map(|entry| {
-                entry.tools.values().filter_map(|tool| {
-                    entry
-                        .exposure_policy
-                        .matches(tool.tool.name.as_ref())
-                        .then(|| tool.clone())
-                })
-            })
-            .collect()
-    }
-
-    pub async fn healthy_tools_for_upstream(&self, upstream: &str) -> Vec<UpstreamTool> {
-        let catalog = self.catalog.read().await;
-        catalog
-            .get(upstream)
-            .into_iter()
-            .filter(|entry| entry.tool_health.is_routable())
-            .flat_map(|entry| {
-                entry.tools.values().filter_map(|tool| {
-                    entry
-                        .exposure_policy
-                        .matches(tool.tool.name.as_ref())
-                        .then(|| tool.clone())
-                })
-            })
-            .collect()
-    }
-
-    async fn has_healthy_tools_for_upstream(&self, upstream: &str) -> bool {
-        let catalog = self.catalog.read().await;
-        catalog.get(upstream).is_some_and(|entry| {
-            entry.tool_health.is_routable()
-                && entry
-                    .tools
-                    .values()
-                    .any(|tool| entry.exposure_policy.matches(tool.tool.name.as_ref()))
-        })
-    }
-
-    pub async fn find_tool_candidates(&self, tool_name: &str) -> Vec<(String, UpstreamTool)> {
-        let catalog = self.catalog.read().await;
-        let mut matches = Vec::new();
-        for (upstream_name, entry) in catalog.iter() {
-            if !entry.tool_health.is_routable() {
-                continue;
-            }
-            if let Some(tool) = entry.tools.get(tool_name)
-                && entry.exposure_policy.matches(tool.tool.name.as_ref())
-            {
-                matches.push((upstream_name.clone(), tool.clone()));
-            }
-        }
-        matches.sort_by(|a, b| a.0.cmp(&b.0));
-        matches
-    }
-
-    pub async fn subject_scoped_tools(
-        &self,
-        configs: &[UpstreamConfig],
-        subject: &str,
-    ) -> Vec<(String, Vec<rmcp::model::Tool>)> {
-        let mut futures = FuturesUnordered::new();
-        let oauth_client_cache = self.oauth_client_cache.clone();
-        for config in configs.iter().filter(|config| config.oauth.is_some()) {
-            let config = config.clone();
-            let subject = subject.to_string();
-            let oauth_client_cache = oauth_client_cache.clone();
-            futures.push(async move {
-                let result = connect_upstream(
-                    &config,
-                    Some(subject.as_str()),
-                    oauth_client_cache.as_ref(),
-                    None,
-                    None,
-                )
-                .await;
-                (config.name.clone(), result)
-            });
-        }
-
-        let mut discovered = Vec::new();
-        while let Some((name, result)) = futures.next().await {
-            match result {
-                Ok((_conn, tools)) => discovered.push((name, tools)),
-                Err(error) => {
-                    tracing::warn!(
-                        upstream = %name,
-                        error = %error,
-                        "subject-scoped upstream tool discovery failed"
-                    );
-                }
-            }
-        }
-        discovered
-    }
-
-    pub async fn subject_scoped_call_tool(
-        &self,
-        config: &UpstreamConfig,
-        subject: &str,
-        params: CallToolRequestParams,
-    ) -> Result<CallToolResult, String> {
-        let start = Instant::now();
-        let tool_name = params.name.to_string();
-        let event = UpstreamRequestLog::tool(&config.name, &tool_name, true)
-            .with_transport(upstream_transport(config));
-        log_upstream_request_start(event);
-        let (conn, _) = match connect_upstream(
-            config,
-            Some(subject),
-            self.oauth_client_cache.as_ref(),
-            None,
-            None,
-        )
-        .await
-        {
-            Ok(conn) => conn,
-            Err(error) => {
-                self.record_failure_for(
-                    &config.name,
-                    UpstreamCapability::Tools,
-                    format!("upstream connect failed: {error}"),
-                )
-                .await;
-                let elapsed_ms = start.elapsed().as_millis();
-                log_upstream_request_error(
-                    event,
-                    elapsed_ms,
-                    "upstream_connect_error",
-                    Some(&error),
-                    None,
-                    None,
-                );
-                return Err(error.to_string());
-            }
-        };
-        match tokio::time::timeout(self.request_timeout, conn.peer.call_tool(params)).await {
-            Ok(Ok(result)) => {
-                let response_size = estimate_response_size(&result);
-                let max_bytes = max_response_bytes();
-                if response_size > max_bytes {
-                    self.record_failure_for(
-                        &config.name,
-                        UpstreamCapability::Tools,
-                        format!("response too large: {response_size} bytes"),
-                    )
-                    .await;
-                    let elapsed_ms = start.elapsed().as_millis();
-                    log_upstream_request_error(
-                        event,
-                        elapsed_ms,
-                        "response_too_large",
-                        None,
-                        Some(response_size),
-                        Some(max_bytes),
-                    );
-                    return Err(format!(
-                        "upstream response too large ({response_size} bytes, max {max_bytes})"
-                    ));
-                }
-                self.record_success_for(&config.name, UpstreamCapability::Tools)
-                    .await;
-                let elapsed_ms = start.elapsed().as_millis();
-                log_upstream_request_finish(event, elapsed_ms, Some(response_size));
-                Ok(result)
-            }
-            Ok(Err(error)) => {
-                self.record_failure_for(
-                    &config.name,
-                    UpstreamCapability::Tools,
-                    format!("upstream call failed: {error}"),
-                )
-                .await;
-                let elapsed_ms = start.elapsed().as_millis();
-                log_upstream_request_error(
-                    event,
-                    elapsed_ms,
-                    "upstream_error",
-                    Some(&error),
-                    None,
-                    None,
-                );
-                Err(format!("upstream call failed: {error}"))
-            }
-            Err(_) => {
-                let message = format!(
-                    "upstream call timed out after {}ms",
-                    self.request_timeout.as_millis()
-                );
-                self.record_failure_for(&config.name, UpstreamCapability::Tools, message.clone())
-                    .await;
-                let elapsed_ms = start.elapsed().as_millis();
-                log_upstream_request_error(event, elapsed_ms, "timeout", None, None, None);
-                Err(message)
-            }
-        }
-    }
-
-    /// Return the names of upstreams currently routable for a capability.
-    pub async fn routable_upstream_names(&self, capability: UpstreamCapability) -> Vec<String> {
-        let catalog = self.catalog.read().await;
-        let mut names: Vec<String> = match capability {
-            UpstreamCapability::Resources => {
-                let resource_names = self.resource_upstreams.read().await;
-                resource_names
-                    .iter()
-                    .filter(|name| {
-                        catalog
-                            .get(*name)
-                            .is_some_and(|entry| entry.health_for(capability).is_routable())
-                    })
-                    .cloned()
-                    .collect()
-            }
-            UpstreamCapability::Tools | UpstreamCapability::Prompts => catalog
-                .iter()
-                .filter(|(_, entry)| entry.health_for(capability).is_routable())
-                .map(|(name, _)| name.clone())
-                .collect(),
-        };
-        names.sort_unstable();
-        names.dedup();
-        names
-    }
-
-    /// Look up which upstream owns a given tool name.
-    #[allow(clippy::significant_drop_tightening)]
-    pub async fn find_tool(&self, tool_name: &str) -> Option<(String, UpstreamTool)> {
-        let catalog = self.catalog.read().await;
-        catalog
-            .values()
-            .filter(|entry| entry.tool_health.is_routable())
-            .find_map(|entry| {
-                entry.tools.get(tool_name).and_then(|tool| {
-                    entry
-                        .exposure_policy
-                        .matches(tool_name)
-                        .then(|| (entry.name.to_string(), tool.clone()))
-                })
-            })
-    }
-
-    /// Get the cached schema for a specific upstream tool.
-    #[allow(clippy::significant_drop_tightening)]
-    pub async fn tool_schema(&self, tool_name: &str) -> Option<Value> {
-        let catalog = self.catalog.read().await;
-        catalog.values().find_map(|entry| {
-            entry.tools.get(tool_name).and_then(|tool| {
-                entry
-                    .exposure_policy
-                    .matches(tool_name)
-                    .then(|| tool.input_schema.clone())
-                    .flatten()
-            })
-        })
-    }
-
-    /// Return all discovered tools for one upstream, including hidden tools and exposure metadata.
-    pub async fn tool_exposure_rows(&self, upstream_name: &str) -> Vec<UpstreamToolExposureRow> {
-        let catalog = self.catalog.read().await;
-        let Some(entry) = catalog.get(upstream_name) else {
-            return Vec::new();
-        };
-
-        let mut rows: Vec<UpstreamToolExposureRow> = entry
-            .tools
-            .values()
-            .map(|tool| {
-                let matched_by = entry.exposure_policy.matched_by(tool.tool.name.as_ref());
-                UpstreamToolExposureRow {
-                    name: tool.tool.name.to_string(),
-                    description: tool
-                        .tool
-                        .description
-                        .as_ref()
-                        .map(ToString::to_string)
-                        .filter(|text| !text.trim().is_empty()),
-                    exposed: matched_by.is_some(),
-                    matched_by,
-                }
-            })
-            .collect();
-        rows.sort_by(|left, right| left.name.cmp(&right.name));
-        rows
-    }
-
-    pub async fn cached_upstream_summary(
-        &self,
-        upstream_name: &str,
-    ) -> Option<UpstreamCachedSummary> {
-        let catalog = self.catalog.read().await;
-        let entry = catalog.get(upstream_name)?;
-        let discovered_tool_count = entry.tools.len();
-        let exposed_tool_count = entry
-            .tools
-            .values()
-            .filter(|tool| entry.exposure_policy.matches(tool.tool.name.as_ref()))
-            .count();
-        let discovered_resource_count = entry.resource_count;
-        let exposed_resource_count = if entry.resource_health.is_routable() {
-            entry.resource_count
-        } else {
-            0
-        };
-        let discovered_prompt_count = entry.prompt_count;
-        let exposed_prompt_count = if entry.prompt_health.is_routable() {
-            entry.prompt_count
-        } else {
-            0
-        };
-
-        Some(UpstreamCachedSummary {
-            discovered_tool_count,
-            exposed_tool_count,
-            discovered_resource_count,
-            exposed_resource_count,
-            discovered_prompt_count,
-            exposed_prompt_count,
-        })
-    }
-
-    pub async fn upstream_runtime_metadata(
-        &self,
-        upstream_name: &str,
-    ) -> Option<UpstreamRuntimeMetadata> {
-        self.connections
-            .read()
-            .await
-            .get(upstream_name)
-            .map(|conn| conn.runtime.clone())
-    }
-
     /// Return cached resource URIs keyed by upstream name (used in catalog snapshots).
     pub async fn cached_upstream_resource_uris(&self) -> Vec<(String, Vec<String>)> {
         let catalog = self.catalog.read().await;
@@ -540,95 +203,6 @@ impl UpstreamPool {
         })
     }
 
-    /// Return the current tool health for one upstream.
-    pub async fn upstream_tool_health(&self, upstream_name: &str) -> Option<UpstreamHealth> {
-        let catalog = self.catalog.read().await;
-        catalog.get(upstream_name).map(|entry| entry.tool_health)
-    }
-
-    /// Call a tool on an upstream server.
-    ///
-    /// Returns `None` if the upstream is not connected or the tool is not found.
-    /// Enforces a response size cap (`LAB_UPSTREAM_MAX_RESPONSE_BYTES`, default 10 MB).
-    ///
-    /// Cap layering by transport:
-    /// - **HTTP non-OAuth**: cap is enforced at the rmcp transport layer by
-    ///   `BodyCappedHttpClient` (see `dispatch/upstream/http_client.rs`) —
-    ///   bytes are checked during streaming, *before* allocation.
-    /// - **stdio**: cap is post-hoc here (rmcp's stdio transport buffers the
-    ///   full JSON response before we see it). The check at the end of this
-    ///   function guards against forwarding oversized payloads but cannot
-    ///   prevent the underlying allocation.
-    /// - **HTTP OAuth**: also post-hoc for now — threading the cap through
-    ///   `OauthClientCache` is tracked as a follow-up.
-    ///
-    /// The post-hoc check below is therefore defense-in-depth for HTTP
-    /// non-OAuth and the primary line of defense for stdio / OAuth.
-    pub async fn call_tool(
-        &self,
-        upstream_name: &str,
-        params: CallToolRequestParams,
-    ) -> Option<Result<CallToolResult, String>> {
-        let start = Instant::now();
-        let tool_name = params.name.to_string();
-        let event = UpstreamRequestLog::tool(upstream_name, &tool_name, false);
-        let peer = self
-            .acquire_peer(upstream_name, UpstreamCapability::Tools, "tool.call")
-            .await?;
-        log_upstream_request_start(event);
-        let result = match tokio::time::timeout(self.request_timeout, peer.call_tool(params)).await
-        {
-            Ok(result) => result.map_err(|e| {
-                let elapsed_ms = start.elapsed().as_millis();
-                log_upstream_request_error(
-                    event,
-                    elapsed_ms,
-                    "upstream_error",
-                    Some(&e),
-                    None,
-                    None,
-                );
-                format!("upstream call failed: {e}")
-            }),
-            Err(_) => {
-                let elapsed_ms = start.elapsed().as_millis();
-                log_upstream_request_error(event, elapsed_ms, "timeout", None, None, None);
-                Err(format!(
-                    "upstream call timed out after {}ms",
-                    self.request_timeout.as_millis()
-                ))
-            }
-        };
-
-        // Enforce response size cap.
-        if let Ok(ref r) = result {
-            let response_size = estimate_response_size(r);
-            let max_bytes = max_response_bytes();
-            if response_size > max_bytes {
-                let elapsed_ms = start.elapsed().as_millis();
-                log_upstream_request_error(
-                    event,
-                    elapsed_ms,
-                    "response_too_large",
-                    None,
-                    Some(response_size),
-                    Some(max_bytes),
-                );
-                return Some(Err(format!(
-                    "upstream response too large ({response_size} bytes, max {max_bytes})"
-                )));
-            }
-            let elapsed_ms = start.elapsed().as_millis();
-            log_upstream_request_finish(event, elapsed_ms, Some(response_size));
-        }
-
-        Some(result)
-    }
-
-    /// Record a failure for an upstream, potentially marking it unhealthy.
-    ///
-    /// After [`CIRCUIT_BREAKER_THRESHOLD`] consecutive failures, the upstream
-    /// is excluded from `list_tools` until a successful re-probe.
     /// Render the synthetic `lab://gateway/servers` document.
     ///
     /// Lists every registered upstream (regardless of health) with the
@@ -1452,7 +1026,7 @@ impl Default for UpstreamPool {
 #[cfg(test)]
 mod tests {
     use super::super::types;
-    use super::super::types::ToolExposurePolicy;
+    use super::super::types::{ToolExposurePolicy, UpstreamTool};
     use super::testsupport::*;
     use super::*;
     use crate::mcp::logging::logging_level_rank;
@@ -1516,19 +1090,6 @@ mod tests {
         let pool = UpstreamPool::new();
         assert!(pool.healthy_tools().await.is_empty());
         assert_eq!(pool.upstream_count().await, 0);
-    }
-
-    #[tokio::test]
-    async fn call_tool_times_out_slow_upstream_response() {
-        let pool = slow_response_pool("slow").await;
-
-        let result = pool
-            .call_tool("slow", CallToolRequestParams::new("slow.tool"))
-            .await
-            .expect("upstream is connected")
-            .expect_err("slow tool call should time out");
-
-        assert!(result.contains("timed out"));
     }
 
     #[tokio::test]
@@ -1706,53 +1267,6 @@ mod tests {
                 "upstream.prompt.two".to_string()
             ]
         );
-    }
-
-    #[tokio::test]
-    async fn hidden_upstream_tools_do_not_appear_in_listings() {
-        let pool = UpstreamPool::new();
-        let upstream_name: Arc<str> = Arc::from("github");
-        let tools = test_upstream_tools(
-            &upstream_name,
-            &["search_repos", "github_create_issue", "delete_repo"],
-        );
-        let mut entry = healthy_in_process_entry(Arc::clone(&upstream_name), tools);
-        entry.exposure_policy =
-            ToolExposurePolicy::from_patterns(vec!["search_repos".into(), "github_*".into()])
-                .expect("policy");
-
-        pool.catalog
-            .write()
-            .await
-            .insert("github".to_string(), entry);
-
-        let names: Vec<String> = pool
-            .healthy_tools()
-            .await
-            .into_iter()
-            .map(|t| t.tool.name.to_string())
-            .collect();
-        assert!(names.contains(&"search_repos".to_string()));
-        assert!(names.contains(&"github_create_issue".to_string()));
-        assert!(!names.contains(&"delete_repo".to_string()));
-    }
-
-    #[tokio::test]
-    async fn hidden_upstream_tools_cannot_be_called_directly() {
-        let pool = UpstreamPool::new();
-        let upstream_name: Arc<str> = Arc::from("github");
-        let tools = test_upstream_tools(&upstream_name, &["search_repos", "delete_repo"]);
-        let mut entry = healthy_in_process_entry(Arc::clone(&upstream_name), tools);
-        entry.exposure_policy =
-            ToolExposurePolicy::from_patterns(vec!["search_repos".into()]).expect("policy");
-
-        pool.catalog
-            .write()
-            .await
-            .insert("github".to_string(), entry);
-
-        assert!(pool.find_tool("search_repos").await.is_some());
-        assert!(pool.find_tool("delete_repo").await.is_none());
     }
 
     #[tokio::test]
